refactor: extract embeddings into ironclaw_embeddings crate
Moves embedding providers (OpenAI, NearAI, Ollama, Bedrock + mock), the
LRU cache, and EmbeddingsConfig data shape into a new
crates/ironclaw_embeddings crate. The binary keeps the small resolver
that reads Settings and validates env-driven base URLs against the SSRF
blocklist. AWS Bedrock deps move into the embeddings + llm crates and
the top-level `bedrock` feature forwards to both via crate-feature
forwarding. No behavior changes; call sites updated to import from the
extracted crate (crate::workspace::Embedding* shims removed).
