+++ crates/ironclaw_embeddings/src/mock.rs
@@ -1,0 +1,105 @@
+//! Deterministic mock embedding provider for tests.
+
+use async_trait::async_trait;
+
+use crate::provider::{EmbeddingError, EmbeddingProvider};
+
+/// A mock embedding provider for testing.
+///
+/// Generates deterministic embeddings based on text hash.
+/// Useful for unit and integration tests.
+pub struct MockEmbeddings {
+    dimension: usize,
+}
+
+impl MockEmbeddings {
+    /// Create a new mock embeddings provider with the given dimension.
+    pub fn new(dimension: usize) -> Self {
+        Self { dimension }
+    }
+}
+
+#[async_trait]
+impl EmbeddingProvider for MockEmbeddings {
+    fn dimension(&self) -> usize {
+        self.dimension
+    }
+
+    fn model_name(&self) -> &str {
+        "mock-embedding"
+    }
+
+    fn max_input_length(&self) -> usize {
+        10_000
+    }
+
+    async fn embed(&self, text: &str) -> Result<Vec<f32>, EmbeddingError> {
+        // Generate a deterministic embedding based on text hash
+        use std::hash::{Hash, Hasher};
+        let mut hasher = std::collections::hash_map::DefaultHasher::new();
+        text.hash(&mut hasher);
+        let hash = hasher.finish();
+
+        let mut embedding = Vec::with_capacity(self.dimension);
+        let mut seed = hash;
+        for _ in 0..self.dimension {
+            // Simple LCG for deterministic random values
+            seed = seed.wrapping_mul(6364136223846793005).wrapping_add(1);
+            let value = (seed as f32 / u64::MAX as f32) * 2.0 - 1.0;
+            embedding.push(value);
+        }
+
+        // Normalize to unit length
+        let magnitude: f32 = embedding.iter().map(|x| x * x).sum::<f32>().sqrt();
+        if magnitude > 0.0 {
+            for x in &mut embedding {
+                *x /= magnitude;
+            }
+        }
+
+        Ok(embedding)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn test_mock_embeddings() {
+        let provider = MockEmbeddings::new(128);
+
+        let embedding = provider.embed("hello world").await.unwrap();
+        assert_eq!(embedding.len(), 128);
+
+        // Check normalization (should be unit vector)
+        let magnitude: f32 = embedding.iter().map(|x| x * x).sum::<f32>().sqrt();
+        assert!((magnitude - 1.0).abs() < 0.001);
+    }
+
+    #[tokio::test]
+    async fn test_mock_embeddings_deterministic() {
+        let provider = MockEmbeddings::new(64);
+
+        let emb1 = provider.embed("test").await.unwrap();
+        let emb2 = provider.embed("test").await.unwrap();
+
+        // Same input should produce same embedding
+        assert_eq!(emb1, emb2);
+    }
+
+    #[tokio::test]
+    async fn test_mock_embeddings_batch() {
+        let provider = MockEmbeddings::new(64);
+
+        let texts = vec!["hello".to_string(), "world".to_string()];
+        let embeddings = provider.embed_batch(&texts).await.unwrap();
+
+        assert_eq!(embeddings.len(), 2);
+        assert_eq!(embeddings[0].len(), 64);
+        assert_eq!(embeddings[1].len(), 64);
+
+        // Different texts should produce different embeddings
+        assert_ne!(embeddings[0], embeddings[1]);
+    }
+}
