+++ crates/ironclaw_embeddings/src/provider.rs
@@ -1,0 +1,57 @@
+//! [`EmbeddingProvider`] trait and shared [`EmbeddingError`] type.
+
+use async_trait::async_trait;
+
+/// Error type for embedding operations.
+#[derive(Debug, thiserror::Error)]
+pub enum EmbeddingError {
+    #[error("HTTP request failed: {0}")]
+    HttpError(String),
+
+    #[error("Invalid response: {0}")]
+    InvalidResponse(String),
+
+    #[error("Rate limited, retry after {retry_after:?}")]
+    RateLimited {
+        retry_after: Option<std::time::Duration>,
+    },
+
+    #[error("Authentication failed")]
+    AuthFailed,
+
+    #[error("Text too long: {length} > {max}")]
+    TextTooLong { length: usize, max: usize },
+}
+
+impl From<reqwest::Error> for EmbeddingError {
+    fn from(e: reqwest::Error) -> Self {
+        EmbeddingError::HttpError(e.to_string())
+    }
+}
+
+/// Trait for embedding providers.
+#[async_trait]
+pub trait EmbeddingProvider: Send + Sync {
+    /// Get the embedding dimension.
+    fn dimension(&self) -> usize;
+
+    /// Get the model name.
+    fn model_name(&self) -> &str;
+
+    /// Maximum input length in characters.
+    fn max_input_length(&self) -> usize;
+
+    /// Generate an embedding for a single text.
+    async fn embed(&self, text: &str) -> Result<Vec<f32>, EmbeddingError>;
+
+    /// Generate embeddings for multiple texts (batched).
+    ///
+    /// Default implementation calls embed() for each text.
+    async fn embed_batch(&self, texts: &[String]) -> Result<Vec<Vec<f32>>, EmbeddingError> {
+        let mut embeddings = Vec::with_capacity(texts.len());
+        for text in texts {
+            embeddings.push(self.embed(text).await?);
+        }
+        Ok(embeddings)
+    }
+}
