+++ src/cli/memory.rs
@@ -7,7 +7,8 @@ use std::sync::Arc;
 
 use clap::Subcommand;
 
-use crate::workspace::{EmbeddingCacheConfig, EmbeddingProvider, SearchConfig, Workspace};
+use crate::workspace::{SearchConfig, Workspace};
+use ironclaw_embeddings::{EmbeddingCacheConfig, EmbeddingProvider};
 
 /// Run a memory command using the Database trait (works with any backend).
 pub async fn run_memory_command_with_db(
