+++ src/config/mod.rs
@@ -57,7 +57,6 @@ pub use self::channels::{
     SignalConfig, TuiChannelConfig,
 };
 pub use self::database::{DatabaseBackend, DatabaseConfig, SslMode, default_libsql_path};
-pub use self::embeddings::{DEFAULT_EMBEDDING_CACHE_SIZE, EmbeddingsConfig};
 pub use self::heartbeat::HeartbeatConfig;
 pub use self::hygiene::HygieneConfig;
 pub use self::llm::default_session_path;
@@ -75,6 +74,7 @@ pub use self::transcription::TranscriptionConfig;
 pub use self::tunnel::TunnelConfig;
 pub use self::wasm::WasmConfig;
 pub use self::workspace::WorkspaceConfig;
+pub use ironclaw_embeddings::{DEFAULT_EMBEDDING_CACHE_SIZE, EmbeddingsConfig};
 // LLM config / session types live in `ironclaw_llm`. Re-exported here so
 // existing `crate::config::*Config` callers (notably `LlmConfig::resolve`
 // in `src/config/llm.rs`, plus the wizard / doctor) keep compiling without
@@ -554,7 +554,7 @@ impl Config {
             owner_id: owner_id.clone(),
             database: DatabaseConfig::resolve()?,
             llm: crate::config::llm::resolve(settings)?,
-            embeddings: EmbeddingsConfig::resolve(settings)?,
+            embeddings: self::embeddings::resolve_embeddings_config(settings)?,
             tunnel,
             channels,
             agent: AgentConfig::resolve(settings)?,
